Add git cloning, refactor
