+++ package.json
@@ -7,8 +7,16 @@
     "src"
   ],
   "scripts": {
-    "lint": "eslint --fix --ignore-path .gitignore .",
+    "lint": "eslint --ignore-path .gitignore .",
     "prepublishOnly": "yarn lint"
+  },
+  "lint-staged": {
+    "*.js": ["eslint --fix --ignore-path .gitignore .", "git add"]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
   },
   "license": "MIT",
   "repository": "github:greenhousecode/create-meh-app",
@@ -19,6 +27,8 @@
     "eslint-config-prettier": "^6.0.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-prettier": "^3.1.0",
+    "husky": "^3.0.4",
+    "lint-staged": "^9.2.3",
     "prettier": "^1.18.2"
   },
   "dependencies": {
