+++ src/config.json
@@ -1,0 +1,4 @@
+{
+  "GITLAB_MEH_NAMESPACE": "GreenhouseGroup/meh",
+  "GITLAB_MEH_NAMESPACE_ID": 4495257
+}
