+++ src/templates/.eslintrc
@@ -1,0 +1,3 @@
+{
+  "extends": ["airbnb-base", "plugin:prettier/recommended"]
+}
