+++ src/templates/.prettierrc
@@ -1,0 +1,5 @@
+{
+  "trailingComma": "all",
+  "singleQuote": true,
+  "printWidth": 100
+}
