+++ src/templates/Dockerfile
@@ -1,0 +1,13 @@
+FROM node:10-alpine
+WORKDIR /usr/src/app
+
+# Install
+COPY package.json yarn.lock ./
+RUN yarn
+
+# Build
+COPY . .
+
+# Serve
+EXPOSE 3000
+CMD ["yarn", "start"]
